requirements.txt
streamlit>=1.24.0
Pillow>=9.0.0
numpy>=1.21.0 